requirements.txt
opencv-python
pillow
numpy
pytesseract
imagehash
requests
pandas
